requirements.txt
mysqlclient
langchain 
openai
python-dotenv